Add support for github action nodes

+++ reboot_ec2_instance.sh
@@ -6,8 +6,14 @@ AWS_SECRET_ACCESS_KEY=$2
 EC2_INSTANCE_ID=$3
 AWS_REGION=$4
 
+if [ -f /usr/bin/aws ]; then
+  echo "AWS cli present no need for installation!"
+else
+  echo "No installation found for aws cli, proceeding with installation..."
+  sudo apt install awscli -y
+fi
 
-sudo apt install awscli -y
+
 aws --profile default configure set aws_access_key_id "$AWS_SECRET_KEY_ID"
 aws --profile default configure set aws_secret_access_key "$AWS_SECRET_ACCESS_KEY"
 aws --profile default configure set region "$AWS_REGION"
